FEAT: #25 찜하기/룰렛추가 기능 구현 완료
두 기능을 모두 구현 완료했다. 두 기능은 detail 페이지에서만
사용되는 기능이 아니기 때문에 common.helpers.js 라는 파일을 만들었다.
앞으로 공통적으로 사용되는 custom hook은 여기에 구현할 예정이다.
찜하기와 룰렛추가 기능이 사용되는 페이지(detail, 둘러보기)에 모두 적용 완료했다.
